Move out common files of `LineChartWidget`
