+++ static/app/views/dashboards/widgets/common/formatChartValue.tsx
