+++ static/app/views/dashboards/widgets/common/shiftTimeserieToNow.tsx
@@ -1,4 +1,4 @@
-import type {TimeseriesData} from '../common/types';
+import type {TimeseriesData} from './types';
 
 export function shiftTimeserieToNow(timeserie: TimeseriesData): TimeseriesData {
   const currentTimestamp = new Date().getTime();
